requirements.txt
google-cloud-bigquery==0.28
